rename: AuthMember 위치 수정
기존에는 dto에 있었는데 vo가 더 적절하다고 생각하여 수정
