+++ backend/src/main/java/com/multi/mariage/auth/controller/AuthController.java
@@ -1,7 +1,7 @@
 package com.multi.mariage.auth.controller;
 
 import com.multi.mariage.auth.annotation.Authenticated;
-import com.multi.mariage.auth.dto.AuthMember;
+import com.multi.mariage.auth.vo.AuthMember;
 import com.multi.mariage.auth.dto.request.LoginRequest;
 import com.multi.mariage.auth.dto.request.ReissueRequest;
 import com.multi.mariage.auth.dto.response.TokenResponse;
