+++ backend/src/main/java/com/multi/mariage/auth/service/AuthService.java
@@ -2,7 +2,7 @@ package com.multi.mariage.auth.service;
 
 import com.multi.mariage.auth.domain.AuthRepository;
 import com.multi.mariage.auth.domain.RefreshToken;
-import com.multi.mariage.auth.dto.AuthMember;
+import com.multi.mariage.auth.vo.AuthMember;
 import com.multi.mariage.auth.dto.request.LoginRequest;
 import com.multi.mariage.auth.dto.request.ReissueRequest;
 import com.multi.mariage.auth.dto.response.TokenResponse;
